// newAPI/src/types/post.ts
import { Types } from "mongoose";
export type Comments = {
  userID: string;
  timestamp: Date;
  content: string;
  _id: Types.ObjectId;
};

type Likes = {
  userID: string;
  timestamp: Date;
};

export type IPost = {
  userID: number;
  timestamp: Date;
  content: string;
  comments: Comments[];
  likes: Likes[];
  _id: Types.ObjectId;
};
export type IPosts = IPost[];